Add files via upload
This is potentially a project that will allow users to view and interact with sporting information like games, their times and statistics utilizing an API. 

User’s would then be able to register and log in on the page and contribute comment’s on games and potentially statistics. User’s may also be able to like or favorite a game as well.